Created components folder and moved app into it
